cogs/utils/context.py
from discord.ext import commands
import discord
import io


class Context(commands.Context):
    def __init__(self, **kwargs):
        super().__init__(**kwargs)

    def __repr__(self):
        # we need this for our cache key strategy
        return '<Context>'

    @property
    def session(self):
        return self.bot.session

    async def safe_send(self, content, *, escape_mentions=True, **kwargs):
        """Same as send except with some safe guards.

        1) If the message is too long then it sends a file with the results instead.
        2) If ``escape_mentions`` is ``True`` then it escapes mentions.
        """
        if escape_mentions:
            content = discord.utils.escape_mentions(content)

        if len(content) > 2000:
            fp = io.BytesIO(content.encode())
            kwargs.pop('file', None)
            return await self.send(file=discord.File(
                fp, filename='message_too_long.txt'),
                                   **kwargs)
        else:
            return await self.send(content)
