requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
discord.py==1.7.2
greenlet==1.0.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.0.1
multidict==5.1.0
munch==2.5.0
Pillow==8.2.0
praw==7.2.0
prawcore==2.0.0
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.13
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.4
websocket-client==0.59.0
yarl==1.6.3
zipp==3.4.1
